filesystem: handle files which change in size correctly
files which grow or shrink are now supported properly
if output mode is file, the header will be patched
according to the actual amount of read data.
This mode also supports reading files from /proc/.
In streaming mode, the size will be fixed at the start of reading the
input. Files which change in size are handled properly here as well,
but due streaming output does not support reading files from /proc/
properly as it depends on a known size before reading which
is not possible with the virtual filesytem in /proc

Signed-off-by: Alexander Mohr <alexander.m.mohr@mercedes-benz.com>